Removed ct_pandoc.cc and set subprocess to be managed in CtProcess
