+++ future/src/ct/CMakeLists.txt
@@ -60,7 +60,6 @@ set(CT_SHARED_FILES
     ct_text_parser.cc
     ct_md_parser.cc
     ct_parser.cc
-    ct_pandoc.cc
     ct_process.cc
     )
